pyproject.toml
[project]
name = "pysignal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "logging-tree>=1.10",
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "rich>=14.1.0",
    "scikit-commpy>=0.8.0",
    "scipy>=1.15.3",
]
